a bit of code cleanup (extra checks for pointers + removed some .h file dependencies)
dvdplayer: some dvds used the wrong subtitle colors that resulted in funny subtitles. Thx to frov for sending me a dvd that had this problem
some subtitles still have color issues when selecting only the .vob file to play

dvdplayer: changed the way the number of subtitles and audiostreams are found.
hopefully it will fix a lot of audio / subtitle issues in the osd (unknown streams and such)

dvdplayer: some debugging utilities


git-svn-id: https://svn.exotica.org.uk:8443/xbmc4xbox@5345 568bbfeb-2a22-0410-94d2-cc84cf5bfa90