changed: define size of audio/video queues based on time instead of bytes
Currently the default is set to 4 seconds which probably is abit on the large side. But I don't really want to add any setting for this as people will think they can solve their crappy networks by increasing this value to some silly amount. A large value will cause GUI delays for certain actions. So increasing it is not a good idea.

git-svn-id: https://svn.exotica.org.uk:8443/xbmc4xbox@25885 568bbfeb-2a22-0410-94d2-cc84cf5bfa90